feat: like 엔티티 추가

+++ src/main/java/kakaotech/communityBE/service/PostService.java
@@ -1,8 +1,11 @@
 package kakaotech.communityBE.service;
 
+import kakaotech.communityBE.dto.PostEditDto;
 import kakaotech.communityBE.dto.PostsDto;
+import kakaotech.communityBE.entity.Like;
 import kakaotech.communityBE.entity.Posts;
 import kakaotech.communityBE.entity.User;
+import kakaotech.communityBE.repository.LikeRepository;
 import kakaotech.communityBE.repository.PostRepository;
 import kakaotech.communityBE.repository.UserRepository;
 import org.slf4j.Logger;
@@ -14,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -22,11 +26,13 @@ public class PostService {
     private static final Logger logger = LoggerFactory.getLogger(PostService.class);
     private final UserRepository userRepository;
     private final PostRepository postRepository;
+    private final LikeRepository likeRepository;
 
     @Autowired
-    public PostService(UserRepository userRepository, PostRepository postRepository) {
+    public PostService(UserRepository userRepository, PostRepository postRepository, LikeRepository likeRepository) {
         this.userRepository = userRepository;
         this.postRepository = postRepository;
+        this.likeRepository = likeRepository;
     }
 
     @Transactional
@@ -58,7 +64,7 @@ public class PostService {
     public List<PostsDto> getPosts() {
         List<Posts> posts = postRepository.findAll();
         List<PostsDto> postsDtos = posts.stream()
-                .map(post -> new PostsDto(post))
+                .map(PostsDto::new)
                 .collect(Collectors.toList());
         return postsDtos;
     }
@@ -66,6 +72,7 @@ public class PostService {
     @Transactional
     public PostsDto getPost(Long postId) {
         Posts posts = getPostsbyId(postId);
+        posts.setViews(posts.getViews() + 1);// 조회수 오르는거 반영
         return new PostsDto(posts);
     }
 
@@ -78,30 +85,40 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long userId, Long postId) {
         Posts posts = getPostsbyId(postId);
-        postRepository.delete(posts);
+        if (userId != posts.getUser().getId()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "게시글 작성자만 글을 지울 수 있습니다.");
+        } else{
+            postRepository.delete(posts);
+        }
     }
 
     @Transactional
-    public PostsDto editPost(Long postId, Long userId, String title, String content, String image){
+    public PostsDto editPost(Long postId, Long userId, PostEditDto postEditDto) {
         Posts posts = getPostsbyId(postId);
         if (!posts.getUser().getId().equals(userId)) {
+            logger.warn("작성자 아님");
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "직성자만 게시글을 수정할 수 있습니다.");
         }
-        posts.setTitle(title);
-        posts.setContent(content);
-        posts.setImage(image);
+        posts.setTitle(postEditDto.getTitle());
+        posts.setContent(posts.getContent());
+        posts.setImage(postEditDto.getImage());
         return new PostsDto(posts);
     }
 
     @Transactional
-    public void viewPost(Long postId) {
-        postRepository.increaseViews(postId);
-    }
-
-    @Transactional
-    public void likePost(Long postId) {
+    public int increaseLikePost(Long postId, Long userId) {
+        User user = getUser(userId);
+        Posts posts = getPostsbyId(postId);
+        Optional<Like> like = likeRepository.findByUserAndPost(user, posts);
+        if (like.isPresent()) {
+            logger.warn("이미 좋아요함");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 좋아요를 눌렀습니다.");
+        }
+        Like newLike = new Like(user, posts);
+        likeRepository.save(newLike);
         postRepository.increaseLikes(postId);
+        return getPostsbyId(postId).getLikes();
     }
 }
